Disable upgrade database prompt.
At present there is no proper upgrade test, it merely checks whether
there are any entries in the database.  As this will be false on first
update, we don't want to have to confirm when we clearly need to perform
an update regardless.

Marked as '#ifdef notyet' so that if in the future the database format
does change we can write a query to detect that situation and re-enable
the check (though arguably we'd always want to upgrade anyway).