Stream pkg_summary downloads via libarchive.
This splits download_summary() into an initialise function (sum_open())
and three archive_read_open() callbacks (sum_start(), sum_read(),
sum_close()).  libarchive can handle EOF and read failures, so this also
simplifies things a little.

One side-effect of this change is that we no longer need individual
progress meters for the download and the database updates (in fact they
would clash), so we now only use the libfetch progress meter.

As we are now updating the database while the file is being fetched, if
the fetch is incomplete we delete the remote summary database rather
than leaving it in an inconsistent state.

With these changes applied on 64-bit SmartOS we see a 2MB reduction in
RSS usage, and a 1.5 second faster runtime for a fresh "pkgin update".