Support PKG_INSTALL_DIR and redo pkg_install setup.
Bring in the latest pkgdb.c from pkg_install and greatly simplify it so
that we only have functions for setting and getting the PKG_DBDIR path.
None of the other database functions were ever used and could have
caused issues depending on if db185 was present.

Add a new setup_pkg_install() which configures pkg_install, and
determine PKG_DBDIR only from pkg_admin, removing the possibility of
defaulting to a compiled-in default that could be completely wrong.

Add support for a PKG_INSTALL_DIR environment variable that allows a
different pkg_install suite to be used.  This is primarily useful for
the pkgin test suite.