Improve pkgin database initialisation.
Fold pkgindb_init(), pkgindb_reset(), pdb_err() and upgrade_database()
into a single pkgindb_open() that correctly opens or creates, checks,
and initialises the database, returning whether the database needs an
update or not.

Now that initialisation is done only when required for a new database,
remove the "IF NOT EXISTS" statements from the database schema, and tidy
up the formatting while here.