Simplify and improve "pkgin stats".
Use the sqlite3_* interfaces directly rather than going via the
pkgindb_doquery() wrapper.  This allows values to be fetched directly
rather than having to be converted, and also enables 4 queries to be
reduced to just one.

Simplify and inline the output rather than having a bunch of defines for
each line.