Some cleanups for remote summary initialisation.
Be verbose when performing an automatic remote summary refresh, the user
is already going to see the download of pkg_summary so we may as well
print the header too for consistency.

Perform a remote refresh for any install operation.  Previously we would
only do this for install, but it needs to be done for upgrades too.

Add and improve comments.