Replace safe_strcmp() with pkgstr_identical().
The semantics of safe_strcmp() were only useful to ensure the two
strings where both non-NULL and identical.  However, often in pkgin we
also allow both strings to be NULL, as not all pkg_summary fields are
mandatory.  The new pkgstr_identical() has clear and safe semantics to
return true iff both are identical (even if both are NULL), otherwise
return false.

Use this new implementation and improve the comment around PKGPATH
checks, as that section could be confusing.

While this does change the semantics of checks for PKGNAME and
FULLPKGNAME, those two shouldn't ever be NULL so we are better served to
ensure that is asserted elsewhere.