Fix refresh upgrades for preserved packages.
Using pkg_add -ffu is not adequate if the version is identical, pkg_add
will not reinstall the package.  We need to use -ffU to force pkg_add to
accept the refresh.

This is merely a stop-gap fix.  Preserve package upgrades are still
handled terribly by ignoring the pkg_delete errors, and will be fixed
properly in due course.