Remove some messages that aren't useful.
The "symlinking ..." message provides no real value, and can confuse
users when upgrades include BUILD_DATE refreshes, as the number of
messages may not match the number of packages to be acted upon if
symlinks already exist from a previous run.

The "installing packages..." message isn't very useful as immediately
after that we print a separate "installing..." message for each package.

The "nothing to upgrade" and "nothing to install" messages are of no use
either, and again can be potentially confusing, for example if all
you've requested is to perform downloads of packages.

While here remove some debug code to make pkgin_install a little more
readable.