Add initial implementation of in-place upgrades.
Instead of performing a remove then reinstall of every package marked
for upgrade, just use "pkg_add -DU" to upgrade in-place.  This has a
number of benefits:

 * Preserved packages are handled correctly and are now upgraded without
   causing errors.
 * Ordering issues no longer cause "package is already installed"
   errors.
 * It is faster.

There are some cases where this isn't optimal yet and can still fail,
for example when trying to handle conflicts where we do legitimately
need to remove packages first before re-installing newer versions.  This
previously wasn't handled correctly either, so it is not considered a
regression, and will be improved in due course.