Stop abusing struct Pkglist for different types.
Get rid of the different types and just have a single Pkglist with a
standard set of entries.  Not only does this simplify things, it avoids
a nasty shadowing issue where Pkglist.old was both an entry and a define
into a different union member.

At some point in the future this will be split up properly into structs
which only have exactly what they need.