Permit the user to install an older pkg version.
For now this counts as an upgrade, even though the version is going
backwards.  In the future this will probably be correctly accounted as a
downgrade action or similar.

Downgrades are only supported for packages that are specifically
requested on the command line, e.g. "pkgin in nodejs-6", otherwise they
are ignored, which was previously the behaviour for all packages.