Show pkg_info output inline instead of buffering.
This avoids an occasional issue where the exec_list() output was being
corrupted.  Buffering the entire output could be considered useful if we
were actually using it to return an error status and print a useful
message, but we weren't, so this way is faster, uses less memory, and
allows us to remove more code.

We also now correctly return the exit code of pkg_info.