Fix handling of unavailable and preferred packages.
During some of the earlier rototilling, the package not available
message was no longer printed for unavailable packages due to changes in
the download code.  This is now fixed, and also differentiates correctly
between packages that are not available, and those that do not match the
requirements stated in preferred.conf.

Use separate functions for install and show-* commands, the former using
the preferred.conf logic, the latter using regular matches to match
local queries and other tools.  This also avoids "not preferred"
messages leaking into show-* output, they belong in trace/debug output.

Abstract away some of the logic that will help in future cleanups, and
improve the output messages.  Upgrades still do not support preferred
matches, this will be fixed in due course.