Ensure single quotes are used for string literals.
SQLite 3.27.0 introduced a change in behaviour and started issuing
warnings to the error log when double quotes were being used for string
literals.  This happens to have been the way pkgin would insert every
package into the database, causing huge disk I/O to the sql.log file
when running "pkgin update".

At best this was significantly impacting runtime by orders of magnitude
and wasting disk space with large sql.log files, and at worst this was
reproducibly causing updates to fail entirely with "Short read of
pkg_summary: Lzma library error:  No progress is possible".

The exact details of why the extra logging was causing lzma failures is
still unclear and needs to be looked into to ensure we aren't missing
something else, but this change is demonstrated to avoid the issue.