Ensure database mtimes are updated correctly.
Previously the database was updated as soon as it was determined that an
update was required, but if the update failed for any reason then pkgin
was left in a broken state thinking it was up-to-date but with no data,
requiring a "pkgin -f update" to fix.

Move the database updates to immediately after the local pkgdb or remote
repository has been inserted.