Ensure upgrades via install are handled correctly.
When an installation results in a package upgrade, we need to consider
the direct reverse dependencies of all upgraded packages, in case there
are shared library changes or similar that could result in the reverse
dependencies no longer functioning correctly.

Users won't always perform an upgrade (which would avoid this as
upgrades always consider all packages) prior to an install.

Resolves issues seen when an installation of a package triggered an
upgrade of openssl but not a refresh of pkgin, resulting in a broken
pkgin binary.