Clean up nbcompat handling.
We shouldn't test for nbcompat headers or library ourselves, pkgsrc
handles pulling it in and setting variables as appropriate depending on
the host system.

Instead, include nbcompat only when in use, otherwise attempt to include
native headers.

At some point it would be nice to ditch nbcompat completely and just
have our own compatibility functions to have a clearer idea of where we
diverge from the standards, and try to reduce where possible.