Switch to automake and overhaul build.
The BSD Makefile was getting crufty, and due to differences between
NetBSD and pkgsrc bootstrap-mk-files it was never consistent with
regards to things like WARNS handling.  There were also ugly hacks
required in pkgsrc to get around it including mk.conf etc.

Take the opportunity to clean up the build process.  All of the
dependencies must now be specified, so that have exactly the correct
library and header paths for each.  Rename some of the options to be
consistent, and update the macros in the code accordingly.

Minor changes include using regular sed instead of finding particular
versions (we use plain flags so have no special requirements), building
and installing the manual page correctly, and removing linking against
libbz2 and libz that have not been used for a long time.