Avoid segfault during upgrade with package removal.
This isn't actually supported at present, but I want some time to create
proper test cases and handle everything correctly, so for now we punt
and ignore removals which is how pkgin behaved prior to 20.7.0.

Apologies for the crashes, but in my defence it exposed a real bug,
which was the reason for removing the checks in the first place.

Should "fix" NetBSDfr/pkgin#94, and reports via private message.