Remove fetchTimeout.
Previously set to 15 seconds, this would kill the connection if no data
was transferred during this time.  There's no particular reason for this
to be set, and is likely the cause of some issues seen in the past with
partial downloads.

Removing the timeout makes pkgin behave like curl's defaults, and also
inadvertently fixes the build issue from NetBSDfr/pkgin#92.