Avoid crashes when remote packages lack FILE_SIZE.
Previously we would dump a warning but then try to continue, eventually
crashing because the remote package URL wasn't set.

Now we warn about the invalid FILE_SIZE and remove the packages from the
available list.  This isn't ideal, as the warning will be shown for all
install-related operations regardless of whether the package has been
selected, but will at least avoid crashing and nag the user to fix the
invalid remote pkg_summary.

Ideally we would want to verify during pkg_summary parsing and avoid the
package being recorded at all, but that will require larger changes to
the pkg_summary parser.