Use savepoints for insert_local/remote_summary.
This way they can be nested inside transactions or other savepoints.

Use savepoint rollback instead of explicitly reversing the actions
that were taken since the savepoint -- simpler and safer this way.
(Actually we could probably just remove this altogether, because the
changes made since the savepoint won't be committed until we
explicitly release the savepoint anyway.)