Remove break_depends().
This hasn't actually done anything since 0.10.0 when with the
introduction of refresh support it effectively ended up being a nop.

Even in older releases, it was of questionable value as it would mark
packages for deletion that were still required, and were only retained
due to pkg_add pulling them in automatically (and that only working if
the user had not run "pkgin clean" before upgrading).

While there are undoubtedly some very rare corner cases where similar
logic is required, this will need to be done correctly as part of a
larger implementation of SUPERSEDES support with correct CONFLICTS
resolution.

As a nice bonus, this improves runtime by around 2x, given we are no
longer performing duplicate dependency calculations.

The pkgin-test-suite now has a test that covers this behaviour, and
shows how 0.9.4 and earlier don't handle it very well.