external: Extend quick_pkg_match to 8 characters.
In pkgin we often match over the entire remote repository, and so it is
optimal to extend this faster check to more characters and reduce the
number of candidate matches, as later pkg_match checks are much more
expensive.

8 loops are currrently chosen as that is the sweet spot for a 2022Q4
repository with around 24,000 packages, limiting the maximum number of
possible matches from over 7000 (thanks to "py") to around 120.

Improves performance of "pkgin upgrade" by around 2x (testing on a
system with around 1200 python packages installed).