Implement LOCAL_REQUIRED_BY table.
This will eventually replace the current method of calculating
LOCAL_REVERSE_DEPS, which currently does not handle all DEPENDS cases.
For example something like p5-DBD-{mysql,mariadb}*-[0-9]* is currently
parsed to have a PKGNAME of "p5-DBD" which is clearly wrong, and this
will ensure accurate reverse dependencies, as even with fixed parsing
of alternative matches it may not be possible to correctly identify
which package is depended upon.

Requires pulling in quite a bit more of pkg_install and duplicating much
of the pkg_admin rebuild-tree functionality which rebuilds +REQUIRED_BY
files.