Clean up and simplify the remove path.
Now that reverse dependencies are calculated correctly based on
REQUIRED_BY a lot of logic can be removed.  Remove simple_pkg_match() as
find_local_pkg_match() can be used to return the correct local pkglist
entry and avoid duplicating strings.

One change in behaviour is that attempts to remove pkg_install are now
handled earlier rather than skipping it during the actual operation, so
that numbers are calculated correctly, and with a better message.