Overhaul install ordering and keep handling.
Package ordering was inadvertently broken since 20.7.0 which ironically
was supposed to fix it.  Packages are now correctly ordered by depth in
all (known) cases.

In addition, dependencies are now installed correctly using "pkg_add -A"
to mark them as automatic, ensuring that the pkgdb is correct at all
times.  Update most of the "keep" path to simplify everything and use
the pkgdb as the repository of record rather than trying to maintain our
own tracking of keep packages during install.

This changes output behaviour with "package marked as non-autoremovable"
messages no longer printed at the end of install operations, but they
were inconsistent anyway and not always printed, and when importing
packages would result in a lot of output of questionable value.

Fixes NetBSDfr/pkgin#128, and should fix NetBSDfr/pkgin#113, with the
new "order" tests in the test suite written to verify correct ordering.