Store file_size in the impact list.
As there are separate pointers to the impact list for download and
install lists we can't use those as do_pkg_install() acts on the install
list but needs to know if the download failed.

Fixes issue when trying to install a package that failed to download.
The http-dl test has been updated to account for this situation.