Implement hashed lookups for the majority of lists.
Add a new array type which is an array of SLISTs, and use them where we
will often look up packages by name.  djb_hash() is used as the hashing
function for good distribution, and each type of hash has its own size
based on performance analysis on the test system.

Many parts of the depends and impact code needed altering to work with
this, and along the way a number of bugs were fixed, mainly around
partial installs where previously we did not correctly consider reverse
dependencies and what happens if they themselves are updated.  Partial
upgrades now correctly recurse both ways for upgraded packages.

As well as fixing bugs, the main benefit of hashed lookups is
performance, and this commit speeds up "pkgin -n upgrade" by 10x on the
test system, and "pkgin -n install zlib" by 5x.