configure.ac: Avoid existence checks in system we're building for.
It is not important that, e.g., /usr/pkg/include/fetch.h exist in the
build environment; it is important that #include <fetch.h> and
linking against -lfetch work.  When cross-building, these files need
to be resolved relative to the destdir, so checking for
/usr/pkg/include/fetch.h is just wrong.  So use AC_CHECK_HEADER,
AC_CHECK_LIB instead of `test -f' and `test -d'.

Allow caller to specify the exact preprocessor flags, linker flags,
and required libraries via LIBxxx_INCLUDES/LDFLAGS/LIBS if needed;
this way if, e.g., libfetch is statically linked and thus requires
the caller to explicitly link with -lssl -lcrypto, the caller can
specify that via `LIBFETCH_LIBS="-lfetch -lssl -lcrypto"'.

This change also renders the `-with-...' options non-mandatory; what
is mandatory is that the header be includable and the library be
linkable.

Note: This will require pkgsrc to build pkgin with:

CONFIGURE_ARGS+=	LIBFETCH_LIBS=${BUILDLINK_LIBS.libfetch}

since otherwise the libfetch detection in configure will choke,
because pkgsrc's libfetch is static.

This should perhaps be adapted to use pkg-config instead but this
simpler change suits my needs, which is to fix the pkgin cross-build.

fix https://github.com/NetBSDfr/pkgin/issues/133