Overhaul CONFLICTS and SUPERSEDES.
Instead of simply rejecting incoming CONFLICTS (or worse, allowing the user
to continue!), if an incoming package conflicts with an installed package
then the installed package is marked for deletion.  This gives a much more
natural interaction when trying to install a package, with the user able to
see what effects will be taken prior to accepting to continue.

At the same time improve SUPERSEDES.  Previously they were only considered
for already-installed packages, and did not handle for example the recent
php-X.Y => phpXY-X.Y overhaul.  Now pkgin will search for the best remote
package that matches the PKGBASE of a package that SUPERSEDES an installed
package.  While not perfect, this is the best we can do with the currently
available pkg_summary(5) fields.