upgrade: Ignore refresh for package tool upgrades.
The reason for performing packaging tool upgrades first is to bring in
fixes or improvements that may be required for the next upgrade.  Having
them upgraded separately first and requiring two runs when they are only
being refreshed quickly gets annoying as, due to pkgsrc velocity, it
happens on pretty much every upgrade when tracking trunk.