autoremove: Ensure lpkg keep is set correctly.
Ordinarily this is set during initialisation based on the pkgin db,
however if the pkgdb changed since the last run and an update was
performed then the in-memory entry wouldn't be set correctly, resulting
in "pkg_add -A" being used for all packages, even though all the other
data was correct.

Fixes issue seen where manual pkg_add (or in fact any pkgdb operation)
prior to an upgrade would result in packages incorrectly being marked as
autoremove.  The test suite has been updated to trigger this case in
https://github.com/TritonDataCenter/pkgin-test-suite/commit/0a61ddc9.